Add files via upload
Keyboard layout for computers running MacOS X following beta code transliteration, since the Greek polytonic keyboard embedded in the system does not follow beta code and is horrible to type. This keyboard is a modification of the Greek Koine RD by Rod Decker. The modification refers to the vowels which receive a Unicode acute GREEK TONOS instead of OXIA according to newest applications on the net, such as Tündra and browser extensions such as Beta Code Greek Converter.  The keyboard was changed using the software Ukelele.